requirements.txt
python-telegram-bot==21.0.1
paho-mqtt==1.6.1
google-cloud-pubsub==2.7.0
pandas==1.4.2
numpy==1.22.3
tensorflow==2.8.0